refactor: restructure repo to avoid circular deps
rust-analyzer does not play well with circular dependencies. See
https://github.com/rust-lang/rust-analyzer/issues/14167 for comments
from many rust-analyzer maintainers around this issue.

Fundamentally, Cargo has (fragile) support for dev-dependencies having
circular dependencies. This works because Cargo makes a distinction
between the integration tests and the actual contract code. This allows
us to not have circular dependencies when building our code, even though
they may exist without an implementation akin to the Cargo build system.

Since rust-analyzer (and likely other tools in the Rust ecosystem) don't
fare well with circular dependencies, we lift the relevant code up to
separate crates (which only contain tests) to ensure there is no
cyclic dependencies.

In the future, we could take a look at removing `white-whale-testing`,
since it does not provide much value (and likely will lead to a higher
chance of circular deps).
