+++ Cargo.toml
@@ -5,9 +5,11 @@ members = [
 	"packages/*",
 	"contracts/liquidity_hub/pool-network/*",
 	"contracts/liquidity_hub/fee_collector",
+	"contracts/liquidity_hub/fee_collector_integration",
 	"contracts/liquidity_hub/fee_distributor",
 	"contracts/liquidity_hub/fee-distributor-mock",
 	"contracts/liquidity_hub/whale_lair",
+	"contracts/liquidity_hub/whale_lair_integration",
 	"contracts/liquidity_hub/vault-network/*",
 	"contracts/liquidity_hub/pool-manager",
 	"contracts/liquidity_hub/epoch-manager",
@@ -23,33 +25,33 @@ repository = "https://github.com/White-Whale-Defi-Platform/migaloo-core"
 publish = false
 
 [workspace.dependencies]
-cosmwasm-schema = { version = "1.1.4"}
+cosmwasm-schema = { version = "1.1.4" }
 cosmwasm-std = { version = "1.1.4", features = ["iterator", "cosmwasm_1_2"] }
-cw2 = { version = "1.0.1"}
-cw20 = { version = "1.0.1"}
+cw2 = { version = "1.0.1" }
+cw20 = { version = "1.0.1" }
 cw20-base = { version = "1.1.0", features = ["library"] }
-cw-storage-plus = { version = "1.1.0"}
-cw-utils = { version = "1.0.1"}
-schemars = { version = "0.8.12"}
-semver = { version = "1.0.12"}
+cw-storage-plus = { version = "1.1.0" }
+cw-utils = { version = "1.0.1" }
+schemars = { version = "0.8.12" }
+semver = { version = "1.0.12" }
 serde = { version = "1.0.145", default-features = false, features = ["derive"] }
-thiserror = { version = "1.0.43"}
-osmosis-std-derive = { version = "0.15.3"}
-prost-types = {version = "0.11.9", default-features = false}
+thiserror = { version = "1.0.43" }
+osmosis-std-derive = { version = "0.15.3" }
+prost-types = { version = "0.11.9", default-features = false }
 white-whale = { path = "./packages/white-whale" }
 white-whale-testing = { path = "./packages/white-whale-testing" }
 cw-multi-test = { version = "0.20.0", features = ["cosmwasm_1_2"] }
-uint = { version = "0.9.5"}
-integer-sqrt = { version = "0.1.5"}
-anyhow = { version = "1.0.71"}
-cw-controllers = { version = "1.1.0"}
-cw-ownable = { version = "0.5.1"}
-anybuf = { version = "0.3.0"}
-sha2 = { version = "=0.10.8"}
-sha256 = { version = "1.4.0"}
+uint = { version = "0.9.5" }
+integer-sqrt = { version = "0.1.5" }
+anyhow = { version = "1.0.71" }
+cw-controllers = { version = "1.1.0" }
+cw-ownable = { version = "0.5.1" }
+anybuf = { version = "0.3.0" }
+sha2 = { version = "=0.10.8" }
+sha256 = { version = "1.4.0" }
 protobuf = { version = "=3.2.0", features = ["with-bytes"] }
 prost = { version = "0.11.9", default-features = false, features = [
-  "prost-derive",
+	"prost-derive",
 ] }
 white-whale-std = { version = "1.1.1", path = "packages/white-whale-std" }
