+++ contracts/liquidity_hub/fee_collector/Cargo.toml
@@ -11,9 +11,9 @@ documentation.workspace = true
 publish.workspace = true
 
 exclude = [
-  # Those files are rust-optimizer artifacts. You might want to commit them for convenience but they should not be part of the source code publication.
-  "contract.wasm",
-  "hash.txt",
+	# Those files are rust-optimizer artifacts. You might want to commit them for convenience but they should not be part of the source code publication.
+	"contract.wasm",
+	"hash.txt",
 ]
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
@@ -41,16 +41,3 @@ serde.workspace = true
 thiserror.workspace = true
 white-whale-std.workspace = true
 cosmwasm-schema.workspace = true
-
-[dev-dependencies]
-cw-multi-test.workspace = true
-terraswap-router = { path = "../pool-network/terraswap_router" }
-terraswap-factory = { path = "../pool-network/terraswap_factory" }
-terraswap-pair = { path = "../pool-network/terraswap_pair" }
-stableswap-3pool = { path = "../pool-network/stableswap_3pool" }
-terraswap-token = { path = "../pool-network/terraswap_token" }
-fee_distributor = { path = "../fee_distributor" }
-whale-lair = { path = "../whale_lair" }
-vault_factory = { version = "1.0.0", path = "../vault-network/vault_factory" }
-vault = { version = "1.0.0", path = "../vault-network/vault" }
-cw20.workspace = true
