+++ contracts/liquidity_hub/fee_collector/src/lib.rs
@@ -5,8 +5,4 @@ mod migrations;
 mod queries;
 pub mod state;
 
-#[cfg(test)]
-#[cfg(not(target_arch = "wasm32"))]
-pub mod tests;
-
 pub use crate::error::ContractError;
