--- contracts/liquidity_hub/fee_collector/src/tests/mod.rs
@@ -1,4 +1,0 @@
-mod common_integration;
-mod dummy_contract;
-mod integration;
-mod testing;
