+++ contracts/liquidity_hub/fee_collector_integration/.gitignore
@@ -1,0 +1,19 @@
+# Generated by Cargo
+# will have compiled files and executables
+/target/
+
+artifacts
+
+# These are backup files generated by rustfmt
+**/*.rs.bk
+
+# macOS
+.DS_Store
+
+# IDEs
+.vscode/
+.idea/
+*.iml
+
+# Auto-gen
+.cargo-ok
