+++ contracts/liquidity_hub/fee_collector_integration/Cargo.toml
@@ -1,0 +1,42 @@
+[package]
+name = "fee_collector_integration"
+version = "1.1.4"
+authors = ["Kerber0x <kerber0x@protonmail.com>"]
+edition.workspace = true
+description = "Integration tests for the fee_collector contract"
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+documentation.workspace = true
+publish.workspace = true
+
+exclude = [
+	# Those files are rust-optimizer artifacts. You might want to commit them for convenience but they should not be part of the source code publication.
+	"contract.wasm",
+	"hash.txt",
+]
+
+[lib]
+path = "lib.rs"
+
+[features]
+# for more explicit tests, cargo test --features=backtraces
+
+[dev-dependencies]
+cosmwasm-std.workspace = true
+cw2.workspace = true
+cw20.workspace = true
+cw-multi-test.workspace = true
+white-whale-std.workspace = true
+fee_collector = { path = "../fee_collector" }
+terraswap-router = { path = "../pool-network/terraswap_router" }
+terraswap-factory = { path = "../pool-network/terraswap_factory" }
+terraswap-pair = { path = "../pool-network/terraswap_pair" }
+stableswap-3pool = { path = "../pool-network/stableswap_3pool" }
+terraswap-token = { path = "../pool-network/terraswap_token" }
+fee_distributor = { path = "../fee_distributor" }
+whale-lair = { path = "../whale_lair" }
+vault_factory = { version = "1.0.0", path = "../vault-network/vault_factory" }
+vault = { version = "1.0.0", path = "../vault-network/vault" }
+serde.workspace = true
+thiserror.workspace = true
