+++ contracts/liquidity_hub/fee_collector_integration/lib.rs
@@ -1,0 +1,5 @@
+// Cargo requires us to have a target specified in the manifest.
+//
+// This is, by default, scanned at `src/lib.rs` and `src/main.rs`, as well as the `[lib]` and `[[bin]]` sections of a `Cargo.toml`.
+//
+// Since we have none, we must manually create a `lib.rs`.
