+++ contracts/liquidity_hub/fee_collector_integration/tests/fee_collector/dummy_contract.rs
