+++ contracts/liquidity_hub/fee_collector_integration/tests/fee_collector/main.rs
@@ -1,0 +1,4 @@
+pub mod common_integration;
+pub mod dummy_contract;
+pub mod integration;
+pub mod testing;
