+++ contracts/liquidity_hub/whale_lair/Cargo.toml
@@ -11,9 +11,9 @@ documentation.workspace = true
 publish.workspace = true
 
 exclude = [
-  # Those files are rust-optimizer artifacts. You might want to commit them for convenience but they should not be part of the source code publication.
-  "contract.wasm",
-  "hash.txt",
+	# Those files are rust-optimizer artifacts. You might want to commit them for convenience but they should not be part of the source code publication.
+	"contract.wasm",
+	"hash.txt",
 ]
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
@@ -39,8 +39,3 @@ semver.workspace = true
 serde.workspace = true
 thiserror.workspace = true
 white-whale-std.workspace = true
-
-[dev-dependencies]
-cw-multi-test.workspace = true
-anyhow.workspace = true
-white-whale-testing.workspace = true
