+++ contracts/liquidity_hub/whale_lair_integration/tests/whale_lair/bond.rs
@@ -3,7 +3,7 @@ use cosmwasm_std::{coin, coins, Decimal, Timestamp, Uint128};
 use white_whale_std::pool_network::asset::{Asset, AssetInfo};
 use white_whale_std::whale_lair::{BondedResponse, BondingWeightResponse};
 
-use crate::tests::robot::TestingRobot;
+use crate::robot::TestingRobot;
 
 #[test]
 fn test_bond_successfully() {
