+++ contracts/liquidity_hub/whale_lair_integration/tests/whale_lair/update_config.rs
@@ -3,7 +3,7 @@ use cosmwasm_std::{Addr, Decimal, Uint128, Uint64};
 use white_whale_std::pool_network::asset::AssetInfo;
 use white_whale_std::whale_lair::Config;
 
-use crate::tests::robot::TestingRobot;
+use crate::robot::TestingRobot;
 
 #[test]
 fn test_update_config_successfully() {
